Configuração do package.json e eslint
