+++ package.json
@@ -1,32 +1,51 @@
 {
-  "name": "md-links",
+  "name": "alinepamplona-md-links",
+  "description": "Extração e verificação de links dentro de arquivos .md",
+  "author": {
+    "name" : "Aline Pamplona",
+    "url" : "https://github.com/alinepamplona"
+  },
   "version": "0.1.0",
+  "keywords": [
+    "node",
+    "markdown",
+    "md",
+    "link",
+    "CLI"
+  ],
+  "homepage": "https://github.com/alinepamplona/SAP010-md-links#readme",
+  "bugs": {
+    "url" : "https://github.com/alinepamplona/SAP010-md-links/issues"
+  },
+  "main": "./src/mdLinks.js",
   "engines": {
     "node": ">=16.x"
   },
   "bin": {
-    "md-links": "./CLI.js"
+    "md-links": "./src/CLI.js"
   },
+  "repository": "github:alinepamplona/SAP010-md-links",
   "bootcamp": {
     "createdAt": "2023-06-19T12:25:44.680Z",
     "version": "6.3.0",
     "commit": "c17498042fc79734cec390e6314cbe810e95bf4c"
   },
   "dependencies": {
-    "axios": "^1.4.0",
-    "eslint": "^8.45.0",
-    "fs": "^0.0.1-security",
-    "jest": "^29.6.1",
     "node": "^20.4.0",
-    "node-fetch": "^3.3.1",
-    "path": "^0.12.7",
-    "test": "^3.3.0"
+    "axios": "^1.4.0",
+    "fs": "^0.0.1-security",
+    "path": "^0.12.7"
   },
   "devDependencies": {
-    "nodemon": "^3.0.1"
+    "eslint": "^8.45.0",
+    "jest": "^29.6.1",
+    "nodemon": "^3.0.1",
+    "test": "^3.3.0"
   },
   "scripts": {
-    "dev": "nodemon CLI.js",
+    "dev": "nodemon ./src/CLI.js",
+    "eslint": "eslint --fix --ext .js src/ test/",
+    "pretest": "npm run eslint",
     "test": "jest --coverage"
   }
 }
